+++ CityApp/src/components/CityStore.tsx
@@ -5,32 +5,38 @@ export interface City {
   id: string;
   name: string;
   country: string;
-  locations?: string[]; 
+  locations?: string[];
 }
 
 const useCityStore = () => {
   const [cities, setCities] = useState<City[]>([]);
   const [loading, setLoading] = useState(true);
 
+  const getCitiesFromStorage = useCallback(async () => {
+    const storedCities = await AsyncStorage.getItem('cities');
+    return storedCities ? JSON.parse(storedCities) : [];
+  }, []);
+
+  const setCitiesToStorage = useCallback((cities: City[]) => {
+    AsyncStorage.setItem('cities', JSON.stringify(cities));
+  }, []);
+
   const fetchCities = useCallback(async () => {
     setLoading(true);
-    const storedCities = await AsyncStorage.getItem('cities');
-    if (storedCities) {
-      const parsedCities = JSON.parse(storedCities);
-      if (Array.isArray(parsedCities)) {
-        setCities(parsedCities);
-      }
+    const storedCities = await getCitiesFromStorage();
+    if (Array.isArray(storedCities)) {
+      setCities(storedCities);
     }
     setLoading(false);
-  }, []);
+  }, [getCitiesFromStorage]);
 
   useEffect(() => {
     fetchCities();
   }, [fetchCities]);
 
   useEffect(() => {
-    AsyncStorage.setItem('cities', JSON.stringify(cities));
-  }, [cities]);
+    setCitiesToStorage(cities);
+  }, [cities, setCitiesToStorage]);
 
   const updateCity = (updatedCity: City) => {
     setCities(currentCities =>
