+++ CityApp/src/components/HomeScreen.tsx
@@ -1,6 +1,6 @@
-import React from 'react';
+import React, { FC, useCallback } from 'react';
 import { FlatList, View } from 'react-native';
-import useCityStore, {City} from './CityStore';
+import useCityStore, { City } from './CityStore';
 import { useNavigation, useFocusEffect } from '@react-navigation/native';
 import { NavigationProp } from '@react-navigation/native';
 import { Button, ActivityIndicator, Card } from 'react-native-paper';
@@ -11,12 +11,24 @@ type RootStackParamList = {
   City: { city: City };
 };
 
-const HomeScreen: React.FC = () => {
+const renderButton = (iconName: string, screen: keyof RootStackParamList, buttonText: string, navigation: NavigationProp<RootStackParamList>) => (
+  <Button icon={iconName} mode="contained" onPress={() => navigation.navigate(screen)}>
+    {buttonText}
+  </Button>
+);
+
+const CityCard: FC<{ item: City, navigation: NavigationProp<RootStackParamList> }> = React.memo(({ item, navigation }) => (
+  <Card onPress={() => navigation.navigate('City', { city: item })}>
+    <Card.Title title={item.name} subtitle={`Country: ${item.country}`} />
+  </Card>
+));
+
+const HomeScreen: FC = () => {
   const navigation = useNavigation<NavigationProp<RootStackParamList>>();
   const { cities, loading, fetchCities } = useCityStore();
 
   useFocusEffect(
-    React.useCallback(() => {
+    useCallback(() => {
       fetchCities();
     }, [fetchCities])
   );
@@ -25,27 +37,16 @@ const HomeScreen: React.FC = () => {
     return <ActivityIndicator animating={true} color="#0000ff" />;
   }
 
-  const renderButton = (iconName: string, screen: keyof RootStackParamList, buttonText: string) => (
-    <Button icon={iconName} mode="contained" onPress={() => navigation.navigate(screen)}>
-      {buttonText}
-    </Button>
-  );
-
   return (
     <View>
-   <View style={{ flexDirection: 'row', justifyContent: 'space-between', margin: 10 }}>
-  {renderButton("information", "Info", "Info")}
-  {renderButton("plus-circle", "AddCity", "Add")}
-</View>
+      <View style={{ flexDirection: 'row', justifyContent: 'space-between', margin: 10 }}>
+        {renderButton("information", "Info", "Info", navigation)}
+        {renderButton("plus-circle", "AddCity", "Add", navigation)}
+      </View>
       <FlatList
         data={cities} 
         keyExtractor={(item) => item.id}
-        renderItem={({ item }) => (
-          <Card onPress={() => navigation.navigate('City', { city: item })}>
-            <Card.Title title={item.name} subtitle={`Country: ${item.country}`} />
-          
-          </Card>
-        )}
+        renderItem={({ item }) => <CityCard item={item} navigation={navigation} />}
       />
     </View>
   );
