+++ CityApp/src/deps.tsx
@@ -1,0 +1,10 @@
+export { default as React, useState, useEffect } from 'react';
+export { NavigationContainer } from '@react-navigation/native';
+export { createStackNavigator, StackNavigationOptions } from '@react-navigation/stack';
+export { default as useCityStore, City } from './components/CityStore';
+export { default as RouteProp } from '@react-navigation/native';
+export { Button, Card, TextInput } from 'react-native-paper';
+import 'react-native-get-random-values';
+export { v4 as uuidv4 } from 'uuid';
+export { useNavigation } from '@react-navigation/native';
+
